fix bug v1.0.14，OrgOwnSealResp.java

+++ easy-esign-struct/src/main/java/io/github/easy/esign/struct/seal/resp/OrgOwnSealResp.java
@@ -12,7 +12,7 @@ public class OrgOwnSealResp {
 
     private String sealName;
 
-    private int sealCreateTime;
+    private Long sealCreateTime;
 
     private boolean defaultSealFlag;
